// web/src/api/generated/model/node.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * stocat
 * OpenAPI spec version: 0.0.1
 */
import type { NodeKind } from "./nodeKind.ts";

export interface Node {
  created_at: string;
  encrypted_name?: string;
  id: string;
  kind: NodeKind;
  library_id: string;
  name?: string;
  name_token?: string;
  parent_id?: string;
  revision: number;
  updated_at: string;
}